Avoid editing an empty buffer on startup
Currently, switching to the worktree after opening Nvim without any path
arguments (without an open file in the current window) results in an
unnamed buffer being created with the path to the current directory
("." is a directory).

With this change, the command to update a currently open file from a
path in a switched worktree is executed only if the buffer in the
current window has a name (saved as a file).